feat: add product and favorites management with handlebars views
- Updated package.json to include express-handlebars for templating.
- Created models for products, favorites, carts, and users.
- Implemented controllers for handling product and favorites logic.
- Added routes for products, favorites, and cart management.
- Developed handlebars views for home, favorites, and cart pages.
- Styled the application with CSS for a better user experience.
- Introduced error handling and validation in controllers.

+++ src/controllers/products.controller.js
@@ -1,4 +1,6 @@
 
+import  productsModel from "../models/products.model.js";
+
 const getProducts = async (req,res)=>{
     try {
         const products = await productsModel.find({})
@@ -12,7 +14,7 @@ const createProduct = async (req,res)=>{
     if (!req.body) {
         return res.status(400).json({status:"error", message: 'Body vacío o mal formado' });
     }
-    const {title,description,price,stock} = req.body
+    const {title,description,price,stock,thumbnail} = req.body
     if(!title || !price || !stock) return res.json({status:"error",message:"faltan datos"})
     try {
         const product = await productsModel.create({title,price,stock,description})
@@ -25,6 +27,7 @@ const createProduct = async (req,res)=>{
 const updateProduct = async (req,res)=>{
 
     const id = req.params.id
+    if(!id) return res.json({status:"error",message:"id invalido"})
     if (!req.body) {
         return res.status(400).json({status:"error", message: 'Body vacío o mal formado' });
     }
@@ -41,10 +44,13 @@ const updateProduct = async (req,res)=>{
 
 const deleteProduct = async (req,res)=>{
     const id = req.params.id
+    if(!id) return res.json({status:"error",message:"id invalido"})
     try {
         const product = await productsModel.findOne({_id:id})
         if(!product) return res.json({status:"error",message:"id invalido"})
         const productDeleted = await productsModel.findByIdAndDelete({_id:id})
+        const productFound = await productsModel.findOne({_id:id})
+        if(productFound) return res.json({status:"error",message:"Error al eliminar el producto"})
         res.json({status:"success",payload:productDeleted})
     } catch (error) {
         res.json({status:"error",message:error.message})
